make use of a baseDevice class
